FIX and TEST Logic

+++ src/main/java/com/virtadity/manease/application/model/product/ProductResponse.java
@@ -1,11 +1,12 @@
 package com.virtadity.manease.application.model.product;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductResponse(
         UUID productId,
         String name,
-        Double weight,
+        BigDecimal weight,
         UUID producerId,
         UUID productTypeId
 ) {
